Added tooltip to screen showing pixel address
When mousing over the screen, a tooltip will appear near the mouse location indicating the memory address of the currently moused-over screen pixel. This tooltip *should* disappear when moving away from the screen, but it is slightly buggy due to the fact that css padding had to be used to trigger the mouse event (as the mouse events only work inside .screen, yet we need to be able to tell it when we're no longer *in* the screen), and the padding can be moused over too quickly or alt-tabbed, causing the tooltip to stay in its last location.